refactor(geo): extract project repair redirect path from geo page
Drops the geo page below React Doctor's control-flow complexity threshold and
gives the repair path its own unit test.
